+++ src/dispatch.py
@@ -1,5 +1,5 @@
 import requests
-from functions.online_ops import  get_latest_news, get_weather_report
+from functions.online_ops import get_weather_report
 from decouple import config
 from datetime import datetime
 from random import choice
@@ -13,7 +13,7 @@ from listener import take_user_input
 from intents import adviceintent, googleintent, jokeintent, myipintent, \
                     synonymintent, whattimeintent, wikipediaintent,\
                     wordlookupintent, youtubeintent, emailintent,\
-                    movieintent
+                    movieintent, newsintent
 from intents.myipintent import find_my_ip
 
 USER=config("USER")
@@ -61,11 +61,8 @@ def start_dispatch(query):
             
 
         elif 'news' in query:
-            play_sound('assets/correct.wav')
-            speak(f"I'm reading out the latest news headlines, sir")
-            speak(get_latest_news())
-            speak("For your convenience, I am printing it on the screen sir.")
-            print(*get_latest_news(), sep='\n')
+            newsintent.handle_intent()
+          
 
         elif 'weather' in query:
             play_sound('assets/correct.wav')
