+++ src/functions/online_ops.py
@@ -13,18 +13,6 @@ EMAIL_PASSWORD = config("EMAIL_PASSWORD")
 
 
 
-
-
-def get_latest_news():
-    news_headlines = []
-    res = requests.get(
-        f"https://newsapi.org/v2/top-headlines?country=in&apiKey={NEWS_API_KEY}&category=general").json()
-    articles = res["articles"]
-    for article in articles:
-        news_headlines.append(article["title"])
-    return news_headlines[:5]
-
-
 def get_weather_report(city):
     url = f"http://api.openweathermap.org/data/2.5/weather?q={city}&appid={OPENWEATHER_APP_ID}&units=imperial"
     res = requests.get(url).json()
